Remove require.extensions support
This is an antipattern in Node, a source of unnecessary complexity
for both Node and CoffeeScript.  It encourages bad behavior and brittle
programming styles.  The Node experiment with JIT compile-to-js on
require() is a failure, and it's high time we all admit that.

Sadly, this breaks all the coffee-script tests, as they only run in
require('x.coffee') mode.